keycloak-plugins/radius-plugin/src/main/java/ua/zaskarius/keycloak/plugins/radius/providers/RadiusProviderSpi.java
package ua.zaskarius.keycloak.plugins.radius.providers;

import org.keycloak.provider.Provider;
import org.keycloak.provider.ProviderFactory;
import org.keycloak.provider.Spi;

public class RadiusProviderSpi implements Spi {

    public static final String RADIUS_SPI = "radius-spi";

    @Override
    public boolean isInternal() {
        return false;
    }

    @Override
    public String getName() {
        return RADIUS_SPI;
    }

    @Override
    public Class<? extends Provider> getProviderClass() {
        return IRadiusConnectionProvider.class;
    }

    @Override
    public Class<? extends ProviderFactory> getProviderFactoryClass() {
        return IRadiusProviderFactory.class;
    }
}
